step: update local-server-start to APIs only mode

+++ templates/local-server-start.sh
@@ -17,14 +17,20 @@ info "Starting project..."
 
 # ==========================================================
 
-cp node_modules/@bgord/design/dist/main.min.css static/
-cp node_modules/@bgord/design/dist/normalize.min.css static/
-info "Copied CSS from @bgord/design"
+if test -f "static/"
+then
+  cp node_modules/@bgord/design/dist/main.min.css static/
+  cp node_modules/@bgord/design/dist/normalize.min.css static/
+  info "Copied CSS from @bgord/design"
+fi
 
 # ==========================================================
 
-./bgord-scripts/frontend-build.sh
-info "Built frontend"
+if test -f "frontend/"
+then
+  ./bgord-scripts/frontend-build.sh
+  info "Built frontend"
+fi
 
 # ==========================================================
